new repository
A project will ask the user for two input numbers and test the input values to determine if the inputs are numbers. If not, the project will notify the user and ask them to enter a new input value until it is a number, then perform multiple mathematical calculations and display the results.